requirements.txt
moviepy==1.0.3
opencv-python==4.8.1.78
Pillow==10.4.0
streamlit==1.31.1
python-dotenv==1.0.1
openai==1.12.0
webvtt-py==0.4.6
openai-whisper==20231117
numpy>=1.21.0 